Bug Fix
When one of the columns was completely filled with tokens, the last token which was slotted in slides all the way down and replaces every circle with a white background. After this, no tokens could be dropped into the column and it was left completely empty.

To fix this issue, I wrapped the setInterval() function within an if statement which checks whether the next token will take the top spot. If it does, don't run the setInterval() function and simply change the color of that circle.